requirements.txt
numpy
Pillow
tqdm